R port of Python SpringRank
This commit adds a fairly straightforward R port of the Python SpringRank
algorithm (refer issue #3). It's basically a line-for-line translation, with few
considerations for "proper" R style. I believe the only difference in
implementation is that Scipy uses an iterative solver, whereas the R
version does not. (This should in principle be a straightforward change but
I know next to nothing about numerical methods, or on what best
practices are when calling something other than simple `solve()` in R.)

This commit also includes an R implementation of the generative model. I
haven't tested it extensively, but (i) it produces correct-looking
graph (in terms of N, average and s.d. of degree) and (ii) it's a fairly
simple model so there shouldn't have been too many places for me to
screw it up!